+++ xquant/keras/dataset_utils.py
@@ -1,0 +1,29 @@
+#  Copyright 2024 Sony Semiconductor Israel, Inc. All rights reserved.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+#  ==============================================================================
+from typing import Any
+
+from xquant.common.dataset_utils import DatasetUtils
+
+
+class KerasDatasetUtils(DatasetUtils):
+
+    @staticmethod
+    def wrapped_dataset(dataset: Any, is_validation: bool, device: str = None) -> Any:
+        def process_data(x: Any, is_validation: bool) -> Any:
+            return x[0] if is_validation else x
+
+        for x in dataset():
+            yield process_data(x, is_validation)
+
